[FIX] Remplacement de l'argent par des bigdecimaux au lieu des doubles

+++ src/main/java/services/CommandService.java
@@ -3,10 +3,12 @@ package services;
 import model.Message;
 import model.Order;
 
+import java.math.BigDecimal;
+
 public interface CommandService {
 
     void sendOrderToDrinkMaker(Order order);
     void sendMessageToDrinkMaker(Message message);
-    void insertMoney(double amount);
+    void insertMoney(BigDecimal amount);
     void printDailyReport();
 }
